Fix ARMv7 build warnings https://bugs.webkit.org/show_bug.cgi?id=288678
Reviewed by Darin Adler.

There are many cast alignment warnings that we skip on 32-bit, mostly related
to GObject* downcasts. Since these warnings aren't very helpful, we just disable
them on 32-bit.

We also fix a few uint64_t <> size_t confusions with format strings.

* Source/WebCore/CMakeLists.txt:
* Source/WebCore/platform/graphics/gstreamer/mse/AppendPipeline.cpp:
* Source/WebCore/platform/graphics/opentype/OpenTypeMathData.cpp:
* Source/WebCore/platform/graphics/texmap/coordinated/CoordinatedBackingStoreTile.cpp:
(WebCore::CoordinatedBackingStoreTile::processPendingUpdates):
* Source/WebCore/platform/libwpe/PlatformPasteboardLibWPE.cpp:
(WebCore::PlatformPasteboard::getTypes const):
(WebCore::PlatformPasteboard::readString const):
* Source/WebKit/UIProcess/API/C/wpe/WKPagePrivateWPE.cpp:
(WKPageHandleKeyboardEvent):
* Source/cmake/WebKitCommon.cmake:
* Tools/MiniBrowser/wpe/main.cpp:
* Tools/Scripts/webkitdirs.pm:
(generateBuildSystemFromCMakeProject):

Canonical link: https://commits.webkit.org/292361@main

+++ Source/WebCore/platform/graphics/gstreamer/mse/AppendPipeline.cpp
@@ -843,7 +843,7 @@ std::pair<AppendPipeline::CreateTrackResult, AppendPipeline::Track*> AppendPipel
     auto preferredTrackId = getStreamIdFromPad(demuxerSrcPad).value_or((static_cast<TrackID>(newTrackIndex)));
     auto trackInfo = m_sourceBufferPrivate.registerTrack(preferredTrackId, streamType);
 
-    GST_DEBUG_OBJECT(pipeline(), "Creating new AppendPipeline::Track with type %s, index %" PRIu64" and id '%" PRIu64 "'", streamTypeToString(streamType), trackInfo.index, trackInfo.id);
+    GST_DEBUG_OBJECT(pipeline(), "Creating new AppendPipeline::Track with type %s, index %" PRIu64" and id '%" PRIu64"'", streamTypeToString(streamType), static_cast<uint64_t>(trackInfo.index), static_cast<uint64_t>(trackInfo.id));
     m_tracks.append(makeUnique<Track>(trackInfo.id, streamType, parsedCaps, presentationSize));
     Track& track = *m_tracks.at(newTrackIndex);
